+++ docs/griptape-framework/misc/src/events_3.py
@@ -1,17 +1,27 @@
-from typing import cast
-
 from griptape.drivers import OpenAiChatPromptDriver
-from griptape.events import CompletionChunkEvent, EventBus, EventListener
+from griptape.events import ActionChunkEvent, EventBus, EventListener, TextChunkEvent
 from griptape.structures import Pipeline
 from griptape.tasks import ToolkitTask
 from griptape.tools import PromptSummaryTool, WebScraperTool
 
+
+def action_chunk_listener(event: ActionChunkEvent) -> None:
+    if event.tag is not None and event.name is not None and event.path is not None:
+        print(f"{event.name}.{event.tag} ({event.path}) ", end="", flush=True)
+    if event.partial_input is not None:
+        print(event.partial_input, end="", flush=True)
+
+
 EventBus.add_event_listeners(
     [
         EventListener(
-            lambda e: print(cast(CompletionChunkEvent, e).token, end="", flush=True),
-            event_types=[CompletionChunkEvent],
-        )
+            lambda e: print(e.token, end="", flush=True),
+            event_types=[TextChunkEvent],
+        ),
+        EventListener(
+            action_chunk_listener,
+            event_types=[ActionChunkEvent],
+        ),
     ]
 )